feat: unlimited implementation of meter attribute getter

+++ src/Cryptlex.LexFloatClient/LexFloatClientNative.cs
@@ -72,10 +72,10 @@ namespace Cryptlex
         public static extern int GetHostLicenseMetadataA(string key, StringBuilder value, int length);
 
         [DllImport(DLL_FILE_NAME, CharSet = CharSet.Unicode, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int GetHostLicenseMeterAttribute(string name, ref uint allowedUses, ref uint totalUses, ref uint grossUses);
+        public static extern int GetHostLicenseMeterAttribute(string name, ref long allowedUses, ref ulong totalUses, ref ulong grossUses);
 
         [DllImport(DLL_FILE_NAME, CharSet = CharSet.Ansi, EntryPoint = "GetHostLicenseMeterAttribute", CallingConvention = CallingConvention.Cdecl)]
-        public static extern int GetHostLicenseMeterAttributeA(string name, ref uint allowedUses, ref uint totalUses, ref uint grossUses);
+        public static extern int GetHostLicenseMeterAttributeA(string name, ref long allowedUses, ref ulong totalUses, ref ulong grossUses);
 
         [DllImport(DLL_FILE_NAME, CharSet = CharSet.Unicode, CallingConvention = CallingConvention.Cdecl)]
         public static extern int GetHostLicenseExpiryDate(ref uint expiryDate);
@@ -142,7 +142,7 @@ namespace Cryptlex
         public static extern int GetHostLicenseMetadata_x86(string key, StringBuilder value, int length);
 
         [DllImport(DLL_FILE_NAME_X86, CharSet = CharSet.Unicode, EntryPoint = "GetHostLicenseMeterAttribute", CallingConvention = CallingConvention.Cdecl)]
-        public static extern int GetHostLicenseMeterAttribute_x86(string name, ref uint allowedUses, ref uint totalUses, ref uint grossUses);
+        public static extern int GetHostLicenseMeterAttribute_x86(string name, ref long allowedUses, ref ulong totalUses, ref ulong grossUses);
 
         [DllImport(DLL_FILE_NAME_X86, CharSet = CharSet.Unicode, EntryPoint = "GetHostLicenseExpiryDate", CallingConvention = CallingConvention.Cdecl)]
         public static extern int GetHostLicenseExpiryDate_x86(ref uint expiryDate);
